Always return VIP as host if set (#179)
## Issue
We would like to always return the configured VIP as the host when
externally connected (to support use cases where the VIP may be
configured outside of the charms e.g. openstack octavia)

## Solution
Update the conditions that 
1. Report host as configured VIP if externally accessible
2. Set active status on the unit if the charm is the leader and VIP is
configured (and prior compatibility checks run - i.e. charm is
externally accessible when related to hacluster + charm has a VIP if
related to hacluster + VIP is configured only when related to
data-integrator)

## Reference
PGBouncer charm has the above described functionality:
https://github.com/canonical/pgbouncer-operator/pull/337/files#diff-b9ed39bbc9c0387bd3e07da31d13373745534a1cd723d3e292c73496b12e307cR551