Revert changes to Allure in #250 (#259)
Reverts part of #250

An Allure Report can only be generated for one of the test suite
combinations (otherwise results with identical test case IDs will
override each other. Any parametrization that pytest is not aware of
[because it was done in spread—i.e. juju version, ubuntu base] will
result in duplicate test case IDs)

Removing the 20.04 base does not change this—we are only able to
generate an Allure Report for one of the Juju versions we test with

TL;DR the Allure Report should only use tests from a single spread
variant. The variant naming was not changed, so the Allure Report config
should not have been changed